requirements.txt
Flask==3.1.2
gunicorn==22.0.0
numpy==2.4.0
pillow==12.1.0
tflite-runtime
